fix new params wal-e reads

+++ core/src/main/java/io/questdb/cairo/wal/WalEventWriter.java
@@ -226,7 +226,7 @@ class WalEventWriter implements Closeable {
         eventMem.putInt(SymbolMapDiffImpl.END_OF_SYMBOL_DIFFS);
     }
 
-    int appendData(long startRowID, long endRowID, long minTimestamp, long maxTimestamp, boolean outOfOrder, long replaceRangeLowTs, long replaceRangeHiTs, short dedupMode) {
+    int appendData(long startRowID, long endRowID, long minTimestamp, long maxTimestamp, boolean outOfOrder, long replaceRangeLowTs, long replaceRangeHiTs, byte dedupMode) {
         startOffset = eventMem.getAppendOffset() - Integer.BYTES;
         eventMem.putLong(txn);
         eventMem.putByte(WalTxnType.DATA);
@@ -238,7 +238,7 @@ class WalEventWriter implements Closeable {
         writeSymbolMapDiffs();
         eventMem.putLong(replaceRangeLowTs);
         eventMem.putLong(replaceRangeHiTs);
-        eventMem.putShort(dedupMode);
+        eventMem.putByte(dedupMode);
         eventMem.putInt(startOffset, (int) (eventMem.getAppendOffset() - startOffset));
         eventMem.putInt(-1);
